Implement a helper class for ontologies
